requirements.txt
absl-py==0.5.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==0.13.0
boto==2.49.0
boto3==1.5.11
botocore==1.8.50
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.6.0
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.2
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.8.2
distributed==1.23.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
funcy==1.13
future==0.16.0
gast==0.3.2
gensim==3.5.0
gevent==1.3.6
glob2==0.7
greenlet==0.4.15
grpcio==1.12.1
h5py==2.8.0
HeapDict==1.0.1
html5lib==1.0.1
hyperopt==0.1.2
idna==2.7
igraph==0.1.11
imageio==2.4.1
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.0
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-numpy==0.4.4.3
multipledispatch==0.6.0
murmurhash==0.28.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==5.7.0
numba==0.39.0
numexpr==2.6.8
numpy==1.14.5
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
openpyxl==2.6.3
packaging==19.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.2.0
pkginfo==1.4.2
plac==0.9.6
pluggy==0.11.0
ply==3.11
preshed==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.6.0
psutil==5.4.7
py==1.8.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.4.2
pyLDAvis==2.1.2
pylint==2.1.1
pymongo==3.7.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.4.2
pypiwin32==223
pyreadline==2.1
PySocks==1.6.8
pytest==3.8.1
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.0
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.0
pywin32==225
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.0.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
regex==2017.11.9
requests==2.13.0
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.1.13
scikit-image==0.14.0
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.8.4
snowballstemmer==2.0.0
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==2.0.12
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
TBB==0.1
tblib==1.4.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.10.3
toolz==0.10.0
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.6
ujson==1.35
umap-learn==0.3.10
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.2.2
xlwings==0.11.8
xlwt==1.3.0
zict==1.0.0